spark-processing/src/main/python/requirements.txt
kafka-python==2.0.2
numpy==1.25.2
